WIP - Refactor crypto
