+++ app/src/main/kotlin/paufregi/connectfeed/core/utils/Crypto.kt
@@ -1,0 +1,56 @@
+package paufregi.connectfeed.core.utils
+
+import android.security.keystore.KeyGenParameterSpec
+import android.security.keystore.KeyProperties
+import java.security.KeyStore
+import javax.crypto.Cipher
+import javax.crypto.KeyGenerator
+import javax.crypto.SecretKey
+import javax.crypto.spec.IvParameterSpec
+
+object Crypto {
+    private const val ALIAS = "ConnectFeedKey"
+    private const val ALGORITHM = KeyProperties.KEY_ALGORITHM_AES
+    private const val BLOCK_MODE = KeyProperties.BLOCK_MODE_CBC
+    private const val PADDING = KeyProperties.ENCRYPTION_PADDING_PKCS7
+    private const val TRANSFORMATION = "$ALGORITHM/$BLOCK_MODE/$PADDING"
+
+    private val cipher = Cipher.getInstance(TRANSFORMATION)
+    private val keyStore = KeyStore
+        .getInstance("AndroidKeyStore")
+        .apply { load(null) }
+
+    private fun getKey(): SecretKey {
+        val existingKey = keyStore.getEntry(ALIAS, null) as? KeyStore.SecretKeyEntry
+        return existingKey?.secretKey ?: createKey()
+    }
+
+    private fun createKey(): SecretKey {
+        return KeyGenerator
+            .getInstance(ALGORITHM).apply {
+                init(
+                    KeyGenParameterSpec.Builder(ALIAS, KeyProperties.PURPOSE_ENCRYPT or KeyProperties.PURPOSE_DECRYPT)
+                        .setBlockModes(BLOCK_MODE)
+                        .setEncryptionPaddings(PADDING)
+                        .setRandomizedEncryptionRequired(true)
+                        .setUserAuthenticationRequired(false)
+                        .build()
+                )
+            }.generateKey()
+    }
+
+    fun encrypt(data: String): ByteArray {
+        cipher.init(Cipher.ENCRYPT_MODE, getKey())
+        val iv = cipher.iv
+        val encrypted = cipher.doFinal(data.encodeToByteArray())
+        return iv + encrypted
+    }
+
+    fun decrypt(bytes: ByteArray): String {
+        val iv = bytes.copyOfRange(0, cipher.blockSize)
+        val data = bytes.copyOfRange(cipher.blockSize, bytes.size)
+        cipher.init(Cipher.DECRYPT_MODE, getKey(), IvParameterSpec(iv))
+        val res = cipher.doFinal(data)
+        return res.decodeToString()
+    }
+}
